Hook up AppRouter with page level components
