+++ client/app/app_router.js
@@ -1,65 +1,33 @@
 /** @jsx React.DOM */
+
 var React = require("react");
-var Promise = require('es6-promise').Promise;
-
-var App = require('./components/app');
-var NotFoundHandler = require('./components/not_found');
-
-
 
 var Router = require("react-router");
 var Routes = Router.Routes;
 var Route = Router.Route;
-var AsyncState = Router.AsyncState;
 var NotFound = Router.NotFound;
 
-var SurveyTake = 
-SurveyEdit = 
-SurveySummary = React.createClass({
-  render: function(){
-    return <div>Hi</div>
-  }
-});
+// Handlers
+var App = require('./components/app');
+var NotFoundHandler = require('./components/not_found');
 
-var SurveyList = React.createClass({
-  mixins:[AsyncState],
-  statics:{
-    getInitialAsyncState: function(path, query, setState){
-      return new Promise(function(resolve, reject){
-        setTimeout(function(){
-          setState({
-            name:'nav'
-          })
-          resolve();
-        }, 1000)
-      });
-    }
-  },
-  render: function(){
-    if(!this.state.name){
-      return <div>Loading ... </div>
-    }
-    return <div>List {this.state.name}</div>
-  }
-});
+var ListSurveys = require('./components/list_surveys');
+var AddSurvey = require('./components/add_survey');
+var EditSurvey = require('./components/edit_survey');
+var SurveySummary = require('./components/survey_summary');
+var TakeSurvey = require('./components/take_survey');
 
-var SurveyAdd = React.createClass({
-  render: function(){
-    return <div>Add</div>
-  }
-});
+var appRouter = (
+  <Routes location="history">
+    <Route title="SurveyBuilder" handler={App}>
+      <Route name="list" path="/" handler={ListSurveys} />
+      <Route title="Add Survey to SurveyBuilder" name="add" path="/add_survey" handler={AddSurvey} />
+      <Route name="edit" path="/surveys/:survey_id/edit" handler={EditSurvey} />
+      <Route name="take" path="/surveys/:survey_id" handler={TakeSurvey} />
+      <Route name="summary" path="/surveys/:survey_id/summary" handler={SurveySummary} />
+      <NotFound title="Page Not Found" handler={NotFoundHandler}/>
+    </Route>
+  </Routes>
+);
 
-var app_router = <Routes location="history">
-      <Route title="SurveyBuilder" handler={App}>
-        <Route name="list" path="/" handler={SurveyList} />
-        <Route title="Add Survey to SurveyBuilder" name="add" path="/add_survey" handler={SurveyAdd} />
-
-        <Route name="edit" path="/surveys/:survey_id/edit" handler={SurveyEdit} />
-        <Route name="take" path="/surveys/:survey_id" handler={SurveyTake} />
-        <Route name="summary" path="/surveys/:survey_id/summary" handler={SurveySummary} />
-        
-        <NotFound title="Page Not Found" handler={NotFoundHandler}/>
-      </Route>
-    </Routes>
-
-module.exports = app_router;
+module.exports = appRouter;
