+++ client/app/components/survey_summary.js
@@ -1,0 +1,11 @@
+/** @jsx React.DOM */
+
+var React = require("react");
+
+var SurveySummary = React.createClass({
+  render: function(){
+    return <div>Survey Summary</div>
+  }
+});
+
+module.exports = SurveySummary;
